% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_bar.R
\name{book_bar}
\alias{book_bar}
\title{Visualization of limit order book}
\usage{
book_bar(
  mbp_order_book,
  implied_book = NULL,
  level = NULL,
  animation = FALSE,
  anim_fps = 10
)
}
\arguments{
\item{mbp_order_book}{Limit order book data generated via `order_book()`,
can be either outright, implied, or consolidated books.}

\item{implied_book}{Implied limit order book generated via `order_book()`
to show the implied liquidity proportion in consolidated book. Default is \code{NULL}
and no implied order book is visualized.}

\item{level}{Number of books depths are visualized. Default is \code{NULL} and
all book depths are visualized.}

\item{animation}{Plot the animated limit order book. Default is \code{FALSE}, and
static plots are returned.}

\item{anim_fps}{Frame per second (FPS) for the animated plot to control the play
speed. Default is \code{10}.}
}
\value{
A list of ggplot2 (static) limit order book plot and an animated limit order book
plot (if \code{animation = TRUE}).
}
\description{
`book_bar()` plots the limit order book with \code{ggplot2::geom_bar()} and an
animated plot is also supported.
}
\section{Data recommendation}{

One would better use resampled data instead of tick-level data to reduce the load
in plotting. Resampling can be done via `cme_resample()`.
}

\section{Animated plot}{

Limit order book data should have multiple timestamps to process the animated plot.
}

\examples{
# This function requires a CME data license to run
# Static plot
\dontrun{
static_plot <- book_bar(book, level = 10, animation = FALSE)

# Animated plot
animation <- book_bar(book, level = 10, animation = TRUE, anim_fps = 40)
}
}
\seealso{
[cme_resample()]
}
