% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cme_resample.R
\name{cme_resample}
\alias{cme_resample}
\title{Resample CME limit order book data}
\usage{
cme_resample(
  mbp_order_book,
  trading.tz = "America/Chicago",
  resample_freq = c("milliseconds", "secs", "mins", "hours"),
  resample_period,
  resample_start,
  resample_end,
  fill = TRUE
)
}
\arguments{
\item{mbp_order_book}{Limit order book data generated via `order_book()`.}

\item{trading.tz}{Time zone of the limit order book and trade summary data.
Default is \code{America/Chicago}.}

\item{resample_freq}{Resample frequency, could be \code{milliseconds}, \code{secs},
\code{mins}, or \code{hours}.}

\item{resample_period}{Resample period (interval), e.g., 1, 10, or 100.}

\item{resample_start}{The starting timestamp of resample data in \code{HH:MM:SS} format.}

\item{resample_end}{The ending timestamp of resample data in \code{HH:MM:SS} format.}

\item{fill}{Fill the gaps if timestamps have irregularities after the resampling. Default is \code{TRUE}.}
}
\value{
A data.table contains the resamped data
}
\description{
Resampling CME limit order book data using \code{highfrequncy::aggregatePrice()}.
}
\examples{
# This function requires a CME data license to run
# 100-millisecond resampling
# Starting timestamp 08:30:00
# Ending timestamp 08:32:00
\dontrun{
resample <- cme_resample(book, trading.tz = "America/Chicago",
resample_freq = "milliseconds", resample_preiod = 100,
resample_start = "08:30:00", resample_end = "08:32:00", fill = TRUE)
}

}
