% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_time.R
\name{heatmap_time}
\alias{heatmap_time}
\title{Visualization of limit order book during a period}
\usage{
heatmap_time(
  mbp_order_book,
  trading.tz = "America/Chicago",
  level = NULL,
  start_timestamp,
  end_timestamp
)
}
\arguments{
\item{mbp_order_book}{Limit order book data generated via `order_book()`,
can be either outright, implied, or consolidated books.}

\item{trading.tz}{Time zone of the limit order book and trade summary data.
Default is \code{America/Chicago}.}

\item{level}{Number of books depths are visualized. Default is \code{NULL} and
all book depths are visualized.}

\item{start_timestamp}{The starting timestamp of the data for the plot in \code{HH:MM:SS} format.}

\item{end_timestamp}{The ending timestamp of the data for the plot in \code{HH:MM:SS} format.}
}
\value{
A ggplot2 figure of limit order book heatmap.
}
\description{
`heatmap_time()` visualizes the limit order book using heatmap plot during a period.
}
\section{Data recommendation}{

One would better use resampled data instead of tick-level data to reduce the load
in plotting. Resampling can be done via `cme_resample()`.
}

\examples{
# This function requires a CME data license to run
# Starting timestamp 08:30:00
# Ending timestamp 08:32:00
\dontrun{
heatmap <- heatmap_time(book, trading.tz = "America/Chicago",
start_timestamp = "08:30:00", end_timestamp = "08:32:00")
}
}
