% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbo_match_details.R
\name{mbo_match_details}
\alias{mbo_match_details}
\title{Extract trade summary with order details from CME Market by Order data}
\usage{
mbo_match_details(mbo_input, price_displayformat = NULL, sunday_input = NULL)
}
\arguments{
\item{mbo_input}{A CME MBO raw data}

\item{price_displayformat}{Display format of trade price, which could be 1,
0.01, or 0.001, etc. If it is unknown, the nearest Sunday's (not a
CME trading holiday) MBP/MBO data need to be provided to extract the display format.}

\item{sunday_input}{When \code{price_displayformat} is not provided, the nearest
Sunday's (not a national holiday)
MBP/MBO data need to be provided to extract the price display format.}
}
\value{
a list of data.table that contains all trade summaries and order details
for all tradeable contracts in a trading day.
}
\description{
`mbo_match_details()` extracts trade summary with details about
matching (limit) orders and an incoming (market) order from CME Market by Order
data, including all fields extracted from `trade_summary()` and order IDs of
the incoming order and matching orders, matched quantity of each matching order,
and trade ID, etc.
}
\section{Trade summary with order details}{

CME Market by Order (MBO) is an order-based data feed that provides details of individual
orders. MBO records every outright order activity at each book depth. Importantly, MBO data show the queue
of individual orders at book depth and relevant information about order submission,
revision, execution and cancellation. Compared to the MBP data, the MBO extends the
market data feed from an aggregated level to ab individual level regarding to each incremental refresh
without losing information covered by the MBP. Different from the MBP data, MBO can show all matching (limit) orders and their corresponding
matching quantity for a trade. Hence, some extra measures related to trades, such as fill rate and execution waiting time, can be
computed.
}

\section{How to read each trade record in the data}{

MBO data assign a unique trade ID for each trade. Generally, most trade records contain at least two order details, i.e., two rows. The
first order detail, i.e., first row indicates the incoming (market) order while the rest order details indicate the matching (limit) orders.
We refer to the CME website and more details can be found at \url{https://cmegroupclientsite.atlassian.net/wiki/spaces/EPICSANDBOX/pages/457225774/MDP+3.0+-+Trade+Summary+Order+Level+Detail}.
\itemize{
For trades with a defined aggressor (agg = 1 or agg = 2):
\itemize{
\item The matched_qty in the first order detail equals the sum of the matched_qty for the rest order details
\itemize{
\item It indicates an incoming market order matches outright orders only.}
\item The matched_qty in the first order detail equals the sum of the matched_qty for rest order details
\itemize{
\item It indicates an incoming market order matches both the outright orders and the implied orders.
The unreported quantity are matched quantity for implied orders.
}
\item When there is only one row for a trade, it probably indicates an incoming (market) order only matches implied orders.
}
For trades without a defined aggressor (agg = 0):
\itemize{
\item For individual contract, it usually happens when an implied out orders are matched in a leg market and the other leg gets matched
as well as a corresponding spread contract, where a trade summary message is sent for the other leg without a defined aggressor.
}
}
}

\section{Fields of trade summary with order details}{

\itemize{
\item Date: Trading date. The night trading session on Sundays is assigned
the trading date as the next Monday.
\item MsgSeq: Message sequence number for each packaet. This is unique for every
message, however, it is not security unique.
\item SendingTime: The timestamp (in UTC) indicates the time when outbound messages to clients
leave the segment gateway (Tag 60).
\item TransactTime: The timestamp (in UTC) indicates the time when incoming messages hit the segment gateway (Tag 52).
\item Code: Contract symbol.
\item Seq: Sequence number, which is security unique. A single packaet may contain information
for multiple tradeable contracts with the single message sequence number, while every
tradeable contract has its unique sequence number.
\item PX: Trade price.
\item Qty: Trade size displayed as the number of contracts.
\item Ord: Number of orders involved in a trade, including both matching (limit) order(s) and an incoming (market) order.
\item agg: Trade aggressor generated by the CME. 1 indicates buyer-initiated trades,
2 indicates seller-initiated trades, while 0 indicates no trade aggressor is defined.
\item trade_id: Unique ID for trades.
\item order_id: Unique ID(s) for incoming (market) order and matching (limit) order(s).
\item matched_qty: The number of contracts that is matched for each individual matching (limit) orders.

}
}

\examples{
# This function requires a CME data license to run
# Example showing how to extract trade summary
# extract trade summary with order details
# Know the display format of trade price
\dontrun{
trade_details <- mbo_match_details(file, "2019-01-07", price_displayformat = 1)

# Unknown price_displayformat
trade_details <- mbo_match_details(file, "2019-01-07", sunday_input = sunday_file)
}

}
\seealso{
[trade_summary()], [meta_data()]
}
