% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbo_quote_queue.R
\name{mbo_quote_queue}
\alias{mbo_quote_queue}
\title{Extract individual order details from CME Market by Order data}
\usage{
mbo_quote_queue(
  mbo_input,
  implied_quotes = FALSE,
  price_displayformat = NULL,
  sunday_input = NULL
)
}
\arguments{
\item{mbo_input}{A CME MBO raw data.}

\item{implied_quotes}{\code{FALSE} by default. if \code{TRUE}, return the implied quote messages.}

\item{price_displayformat}{Display format of trade price, which could be 1,
0.01, or 0.001, etc. If it is unknown, the nearest Sunday's (not a
CME trading holiday) MBP/MBO data need to be provided to extract the display format.}

\item{sunday_input}{When \code{price_displayformat} is not provided, the nearest
Sunday's (not a national holiday)
MBP/MBO data need to be provided to extract the price display format.}
}
\value{
a list of data.table that contains all individual order details
for all tradeable contracts in a trading day.
}
\description{
`mbo_quote_queue()` extracts individual order details for all outright limit orders
 from CME Market by Order data, including all fields extracted from `quote_messages()`
 and order IDs, order priority, etc.
}
\section{Individual order details}{

MBO data record each order with a unique order ID though the order may be revised,
cancelled, filled, or discarded. The life of an order is tracked by the unique order ID. For each
update other than execution, MBO data assign order priority number, from lowest to highest
values, to position the order against other orders of the same side and price. According to
CME Globex rulebook, an order loses its original order priority if the order quantity increases
while the original order priority can be saved if the order quantity decreases. A lower value
of order priority number means a higher priority. Any revision of a limit order, including
price, and quantity will be recorded into MBO with the unique order ID. However, MBO data
do not flag what type of a limit order is, for instance, if an order is day order that will be
discarded automatically by CME if it does not conclude per se, or a GTC (Good till Cancel)
order that can rest in the LOB until this order will be filled. Similarly, the MBO does not
flag stop orders, a FOK (Fill or Kill) order, or a FAK (Fill and Kill) order per se. Market
orders are also assigned with unique order IDs and order priorities, which could be checked
at the first row of a trade record in general.Not all MBO messages are assigned unique sequence numbers
like the MBP but all messages are assigned message numbers though they may not be unique in a few
scenarios.
}

\section{Quote messages---MBO vs MBP}{

CME generally uses two templates for disseminating the individual order details.
MBO update may correspond with an update of MBP. When it happens, the MBO data
disseminate this scenario with dual updates in a message under a combined template, which
indicates that the change in MBP is driven by the update in MBO. This message can clarify
the reason why the limit order book is updated from an order perspective. However,
an incremental refresh of MBO may not result
in an update of MBP since the order submitted at a price that where MBP cannot record.
This could happen in the following scenarios: 1) the depth is behind the best 10th level;
2) A cluster of (or massive) order activities occur in a short period of time and MBP only
update their joint update; 3) When a submitted limit order that improves the best-bid-offer
price in book, the MBP may not be updated simultaneously but with certain time delay.
In this scenario, different message template is used. In addition, if an update of MBP is
not caused by an MBO update, the combined template will be applied but with asymmetric
refreshes. This scenario may take place when market orders consume all liquidity at one or
multiple depths and the rest depths need to be moved forward. This moving process is sent
by updating messages of MBP though no specific order activities are pertinent to this.
}

\examples{
# This function requires a CME data license to run
# Example showing how to extract individual order details
# Know the display format of trade price
\dontrun{
order_details <- mbo_quote_queue(file, "2019-01-07", price_displayformat = 1)

# Unknown price_displayformat
order_details <- mbo_quote_queue(file, "2019-01-07", sunday_input = sunday_file)
}

}
\seealso{
[quote_messages()], [meta_data()]
}
