% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_data.R
\name{meta_data}
\alias{meta_data}
\title{Extract necessary metadata from the CME Market by Price and/or Market by Order data}
\usage{
meta_data(sunday_input, date, security = NULL)
}
\arguments{
\item{sunday_input}{A Sunday CME MBO/MBP raw data.}

\item{date}{Trading date associated with the raw data, which can be found in
the file name easily in \code{YYYYMMDD} format. This argument is required to select
the parsing format.}

\item{security}{A symbol of a specific contract of a financial security, should be
in the format like "CLK9" (crude oil), "ESH9" (E-mini S&P 500), etc. These can
be obtained from CME website for a contract specification.}
}
\value{
a data.table contains the meta data of either a specific contract, or
all tradeable contracts of a financial security traded in the CME.
}
\description{
`meta_data()` extracts necessary security definition of a specific contract, or
all tradeable contract of a financial security, traded in the CME, including security
symbol, maturity, security type (e.g,. futures/options),
exchange (e.g., CME, NYMEX), matching algorithm, minimum price increment (tick size),
price display format, number of market depth supported, implied eligibility, etc.
}
\examples{
# This function requires a CME data license to run
# Example showing how to extract meta data
# extract meta data
# Know the specific security (e,g., ESH9)
\dontrun{
es.meta.data <- meta_data(file, "2019-01-07", 'ESH9')

# For all tradable contracts in E-mini S&P 500 futures
es.meta.data <- meta_data(file, "2019-01-07")
}
}
