% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_book.R
\name{order_book}
\alias{order_book}
\title{Reconstruct limit order book(s) from the quote messages extracted from the CME Market by Price data}
\usage{
order_book(
  mdp_quote_msgs_list,
  level = NULL,
  consolidate = TRUE,
  sunday_input = NULL,
  security = NULL
)
}
\arguments{
\item{mdp_quote_msgs_list}{A list of quote messages extracted from the MBP data using `quote_messages()`, which must \bold{include} Sunday's file.}

\item{level}{The maximum book depth that can be supported by the CME, which could be 5 or 10 for futures, 3 for options.
When it is not provided, the nearest Sunday's (not a national holiday)
MBP/MBO data need to be provided to extract the default depth.}

\item{consolidate}{If \code{TRUE} (default), return the consolidated order book.}

\item{sunday_input}{When \code{level} is not provided, the nearest Sunday's (not a national holiday)
MBP/MBO data need to be provided to extract the default depth.}

\item{security}{A symbol of a single contract of a financial security or a vector of multiple single contracts, should be
in the format like "CLK9" (crude oil), "ESH9" (E-mini S&P 500), etc. These can
be obtained from CME website for a contract specification.}
}
\value{
A list of outright, implied, and consolidated (if \code{consolidate} = \code{TRUE})
limit order books for all tradeable contracts in a trading day stored in a list.
}
\description{
`order_book()` reconstructs limit order book(s) from the quote messages extracted from the CME Market by Price data.
This function can process outright limit order book that is reconstructed by all outright quotes, implied order book
that is reconstructed by all implied quotes, or consolidated order book that aggregates information of both outright
and implied quotes. Outright order book can be supported up to 10 depths, while implied order book can be supported up
to 2 depths, according the CME definition.
}
\section{Why \code{mdp_quote_msgs_list} needs Sunday's file}{

One should store all quote messages extracted from the MBP data including Sunday's file
in a list because CME disseminates the initial snapshot of the book on Sunday. To ensure the
accuracy of book reconstruction, the Sunday's file is needed.
}

\section{Processing specific contract(s)}{

In practice, users might not need to process the limit order books of all tradable contracts,
which is substantially time-consuming. This function provides a way to only reconstruct the
limit order book of a single or multiple trade contracts.
}

\section{Limit order book reconstruction}{

Outright limit order book and implied limit order book are reconstructed separately at first
and they should be consolidated finally according to book depths. For the addition update,
it indicates a better price than a current book depth enters the market and one should put the previous depths backward,
and fill up to ten depths of each side. When a message updated with modification action, one only needs to update the quantity
and number of orders with respect to a book depth, while no move occurs. In terms of cancellation,
it indicates a current depth will be deleted in a side of the book. It also moves the rest depths forward and fill the first ten depths.
Implied limit order book could be constructed with the same logic though the number of orders for each level is not defined in the MBP.
The consolidated LOB can be reconstructed according to timestamps or sequence numbers chronologically.
If a book depth that exists in the outright book, the quantity of implied book is added to the outright book to reconstruct the
consolidated book. When a book depth of implied book is better than the outright book, this depth for the consolidated LOB is the one at implied book
and one should move book depths behind this level in the outright LOB backward. Consolidated book is the one that traders monitor during trading.
}

\section{Level of limit order book}{

One should reconstruct the limit order book with the maximum book depth stipulated by the CME to avoid possible errors.
}

\examples{
# This function requires a CME data license to run
# Example showing how to reconstruct the limit order book(s)
# Know book depth
\dontrun{
book <- order_book(weekly_msg_quotes_list, "2019-01-07", level = 10)

# Unknown book depth
book <- order_book(weekly_msg_quotes_list, "2019-01-07", level = 10, sunday_input = sunday_file)
}
}
\seealso{
[quote_message()]
}
