% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote_messages.R
\name{quote_message}
\alias{quote_message}
\title{Extract quote messages from CME Market by Price data}
\usage{
quote_message(mbp_input, date, price_displayformat = NULL, sunday_input = NULL)
}
\arguments{
\item{mbp_input}{Market by Price raw data.}

\item{date}{Trading date associated with the raw data, which can be found in
the file name easily in \code{YYYYMMDD} format. This argument is required to select
the parsing format.}

\item{price_displayformat}{Display format of trade price, which could be 1,
0.01, or 0.001, etc. If it is unknown, the nearest Sunday's (not a CME trading holiday)
MBP/MBO data need to be provided to extract the display format.}

\item{sunday_input}{When \code{price_displayformat} is not provided, the nearest
Sunday's (not a CME trading holiday) MBP/MBO data needs to be provided to extract
the price display format.}
}
\value{
a list of data.table that contains all quote messages (including
implied quote messages if they are available) for all tradeable contracts in a trading
day.
}
\description{
`quote_message()` extracts all quotes messages based on price levels in a
limit order from the whole CME Market by Price (MBP) data, including
timestamps, quote update (add/modify/cancel),
quote price, quote side (bid or offer), quote depth, quote quantity,
number of orders in a price depth, etc. Implied quote messages are also
considered if they are available in a market.
}
\section{Quote messages}{

Quote messages include both outright quotes and implied quotes. Outright quotes
are those submitted by real traders while the implied quotes are those generated through
other markets. In terms of outright quotes, each update defines how the outright limit order book updates for a book depth.
Quote messages only show updates to the book up to ten depths. The quoted quantity
and the number of orders resting in each depth can be defined for outright orders exclusively. For
the addition, it indicates a better price than a current quote entries the market.
When a message updated with modification, it indicates the quantity of an existing depth
is changed.In terms of update of cancellation, it indicates a current depth is deleted in the book.
Similarly, implied quote messages show updates for the implied limit order book and only two depths are dissminated.
}

\section{Fields of quote messages}{

\itemize{
\item Date: Trading date. The night trading session on Sundays is assigned
the trading date as the next Monday.
\item MsgSeq: Message sequence number for each packaet. This is unique for every
message, however, it is not security unique.
\item SendingTime: The timestamp (in UTC) indicates the time when outbound messages to clients
leave the segment gateway (Tag 60).
\item TransactTime: The timestamp (in UTC) indicates the time when incoming messages hit the segment gateway (Tag 52).
\item Code: Contract symbol.
\item Seq: Sequence number, which is security unique. A single packaet may contain information
for multiple tradeable contracts with the single message sequence number, while every
tradeable contract has its unique sequence number.
\item Update: Indicate whether the quote is for addition (0), modification (1), or cancellation (2).
\item Side: A quote is on either bid side (0 or E) or ask side (1 or F).
\item PX: Quoted price.
\item Size: Quote size displayed as the number of contracts for a depth in the book.
\item Ord: Number of limit orders joining the queue for a depth in the book.
\item Implied: Whether a quote is for either implied orders (Y) or outright orders (N).
\item PX_depth: Book depth where a quote update occurs.
For example, 1 indicates the BBO and 2 indicates the second best depth, etc. Outright books
support up to 10 depths, and implied books support up to 2 depths.
}
}

\examples{
# This function requires a CME data license to run
# Example showing how to extract quote messages
# extract quote messages
# Know the display format of quote price
\dontrun{
quotes <- quote_message(file, "2019-01-07", price_displayformat = 1)

# Unknown price_displayformat
quotes <- quote_message(file, "2019-01-07", sunday_mbp_input = sunday_file)
}

}
\seealso{
[meta_data()]
}
