% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\title{Extract trading status from the CME Market by Price and/or Market by Order data}
\usage{
status(input, date)
}
\arguments{
\item{input}{CME MBO/MBP data}

\item{date}{Trading date associated with the raw data, which can be found in
the file name easily in \code{YYYYMMDD} format. This argument is required to select
the parsing format.}
}
\value{
A data.table contains sequence numbers, trading session indicator, etc.
}
\description{
`status()` extracts the trading status from the CME Market by Price and/or Market
by Order data. This function indicates the sequence numbers
of pre-open session, opening session, continuous trading session, etc.
}
\examples{
# This function requires a CME data license to run
# Example showing how to extract trading status
# Know the specific security (e,g., ESH9)
\dontrun{
es.status <- status(file, "2019-01-07")

# For all tradable contracts in E-mini S&P 500 futures
eg.status <- status(file, "2019-01-07")
}

}
