% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbbo.R
\name{tbbo}
\alias{tbbo}
\title{Build the TBBO data from the trades and quotes data}
\usage{
tbbo(
  trades,
  book,
  merge = c("TransactTime", "Seq_number"),
  assign_trades = FALSE
)
}
\arguments{
\item{trades}{Trade summary data generated from `trade_summary()`.}

\item{book}{Limit order book data generated from `order_book()`.}

\item{merge}{The method to merge the trade and quote data, which could be either
Transaction Time (\code{TransacTime}) or Sequence Number (\code{Seq_number}).}

\item{assign_trades}{Assess trade direction based on Lee and Ready (1991), including
both quote test and tick test}
}
\value{
A data.table contains the trades and quotes infomation immediately prior to the trades
}
\description{
`tbbo()` builds the TBBO data where records the trade summary with best-bid-offer
information immediately prior to trades happening. This function also provides
Lee and Ready (1991)'s trade assignment assessment.
}
\examples{
# This function requires a CME data license to run
# Example showing how to build the TBBO
# Merge by Transaction Time and assess trade direction
\dontrun{
TBBO <- tbbo(trades, book, merge = "TransactTime", assign_trades = TRUE)

# Merge by Sequence Number and assess trade direction
TBBO <- tbbo(trades, book, merge = "Seq_number", assign_trades = TRUE)
}

}
\references{
Charles M. C. Lee, and Mark J. Ready. 1991. “Inferring Trade Direction from Intraday Data.” Journal of Finance 46, 733–46.
}
