% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tick_bbo.R
\name{tick_bbo}
\alias{tick_bbo}
\title{Visualization of trade price and best-bid-offer prices}
\usage{
tick_bbo(
  mbp_order_book,
  trade_summary,
  trading.tz = "America/Chicago",
  start_timestamp,
  end_timestamp
)
}
\arguments{
\item{mbp_order_book}{Limit order book data generated via `order_book()`.}

\item{trade_summary}{Trade summary data generated via `trade_summary()`.}

\item{trading.tz}{Time zone of the limit order book and trade summary data.
Default is \code{America/Chicago}.}

\item{start_timestamp}{The starting timestamp of the data for the plot in \code{HH:MM:SS} format.}

\item{end_timestamp}{The ending timestamp of the data for the plot in \code{HH:MM:SS} format.}
}
\value{
A ggplot2 figure of trade price and best-bid-offer prices.
}
\description{
`tick_bbo()` plots the trade price and best-bid-offer prices during a period.
}
\examples{
# This function requires a CME data license to run
# Starting timestamp 08:30:00
# Ending timestamp 08:32:00
\dontrun{
tick_plot <- tick_bbo(book, trade, trading.tz = "America/Chicago",
start_timestamp = "08:30:00", end_timestamp = "08:32:00")
}
}
