% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{statistics}
\alias{statistics}
\title{Extract daily and session statistics from the CME Market by Price and/or Market by Order data}
\usage{
statistics(input, date)
}
\arguments{
\item{input}{CME MBO/MBP data}

\item{date}{Trading date associated with the raw data, which can be found in
the file name easily in \code{YYYYMMDD} format. This argument is required to select
the parsing format.}
}
\value{
a data.table contains the statistics of all tradeable contracts of a security in the CME.
}
\description{
`statistics()` extracts daily and session statistics from the CME Market by Price
and/or Market by Order data, including indicative (open) price, settlement price,
open interest, trading volume, low price, high price, highest bid, lowest offer, etc.
}
\section{Supported statistics}{

\itemize{
\item open: Indicative open and open prices
\item settle: Settlement prices
\item high_px: Session high trade Price
\item low_px: Session low trade price
\item low_offer: Session lowest offer price
\item high_bid: Session highest bid price
\item volume: Daily trading volume
\item open_int: Daily open interest
\item elec_volume: Number of contracts traded in a session
\item limit: Limit high and low price
\item simulate_buy (only supported prior to Nov 20, 2015): Pre-open simulated buy price
\item simulate_sell (only supported prior to Nov 20, 2015): Pre-open simulated sell price
}
}

\examples{
# This function requires a CME data license to run
# Example showing how to extract daily and session statistics
\dontrun{
All tradable contracts in E-mini S&P 500 futures
es.stat <- statistics(file, "2019-01-07")
}

}
